+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.0.18 (Unreleased)
+~~~~~~~~~~~~~~~~~~~
+
+-  -  :class:`~ocdsextensionregistry.extension_version.ExtensionVersion`: :meth:`~ocdsextensionregistry.extension_version.ExtensionVersion.__repr__` falls back to Base URL and Download URL if Id or Version is blank.
+
+
 0.0.17 (2020-04-03)
 ~~~~~~~~~~~~~~~~~~~